+++ picodi/helpers.py
@@ -5,6 +5,7 @@ import contextlib
 from typing import TYPE_CHECKING, Any, ParamSpec, TypeVar, overload
 
 import picodi
+from picodi import ManualScope
 
 if TYPE_CHECKING:
     from collections.abc import AsyncGenerator, Callable, Generator
@@ -116,9 +117,7 @@ class _Lifespan:
     @contextlib.contextmanager
     def sync(
         self,
-        scope_class: (
-            type[picodi.ScopeType] | tuple[type[picodi.ScopeType]]
-        ) = picodi.ManualScope,
+        scope_class: type[ManualScope] | tuple[type[ManualScope]] = ManualScope,
     ) -> Generator[None, None, None]:
         picodi.init_dependencies(scope_class)
         try:
@@ -129,9 +128,7 @@ class _Lifespan:
     @contextlib.asynccontextmanager
     async def async_(
         self,
-        scope_class: (
-            type[picodi.ScopeType] | tuple[type[picodi.ScopeType]]
-        ) = picodi.ManualScope,
+        scope_class: type[ManualScope] | tuple[type[ManualScope]] = ManualScope,
     ) -> AsyncGenerator[None, None]:
         await picodi.init_dependencies(scope_class)
         try: